# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "OnLaunch - API Documentation",
    "version": "1.0"
  },
  "paths": {
...
    "/api/admin/v0.1/app/message/{messageId}": {
      "get": {
        "tags": [
          "Admin API"
        ],
        "summary": "Retrieve a message by ID",
        "description": "Retrieves a message along with its actions for a given message ID.",
        "parameters": [
          {
            "in": "path",
            "name": "messageId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The message ID"
          }
        ],
        "responses": {
          "200": {
            "description": "A message object",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          },
          "404": {
            "description": "No message found with the provided ID"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin API"
        ],
        "summary": "Delete a message by ID",
        "description": "Deletes a message for a given message ID.",
        "parameters": [
          {
            "in": "path",
            "name": "messageId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The message ID to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          },
          "404": {
            "description": "No message found with the provided ID"
          }
        }
      },
      "put": {
        "tags": [
          "Admin API"
        ],
        "summary": "Update a message",
        "description": "Updates an existing message with new information.",
        "parameters": [
          {
            "in": "path",
            "name": "messageId",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "The message ID to update"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully updated the message",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          },
          "404": {
            "description": "No message found with the provided ID"
          }
        }
      }
    },
    "/api/admin/v0.1/app/messages": {
      "get": {
        "tags": [
          "Admin API"
        ],
        "summary": "Get all messages for an app",
        "description": "Retrieves all messages associated with an app, identified by the app's authentication credentials.",
        "responses": {
          "200": {
            "description": "An array of message objects",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MessageDto"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed, invalid or missing credentials"
          }
        }
      },
      "post": {
        "tags": [
          "Admin API"
        ],
        "summary": "Create a new message",
        "description": "Creates a new message for an app, with details provided in the request body.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessageDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Message created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MessageDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed for the provided message data"
          },
          "401": {
            "description": "Authentication failed, invalid or missing credentials"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "MessageDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "blocking": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActionDto"
            }
          }
        }
      },
...
      "ActionDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "actionType": {
            "type": "string",
            "enum": [
              "DISMISS",
              "OPEN_APP_IN_APP_STORE"
            ]
          },
          "buttonDesign": {
            "type": "string",
            "enum": [
              "FILLED",
              "TEXT"
            ]
          }
        }
      },
      "CreateMessageDto": {
        "type": "object",
        "required": [
          "blocking",
          "title",
          "body",
          "startDate",
          "endDate"
        ],
        "properties": {
          "blocking": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date-time"
          },
          "endDate": {
            "type": "string",
            "format": "date-time"
          },
          "actions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActionDto"
            }
          }
        }
      },
...
    }
  },
...
}